test fix location of encode bedrock runtime modelid arn
