+++ litellm/passthrough/main.py
@@ -24,8 +24,8 @@ import litellm
 from litellm.litellm_core_utils.get_llm_provider_logic import get_llm_provider
 from litellm.llms.custom_httpx.http_handler import AsyncHTTPHandler, HTTPHandler
 from litellm.llms.custom_httpx.llm_http_handler import BaseLLMHTTPHandler
+from litellm.passthrough.utils import CommonUtils
 from litellm.utils import client
-from litellm.proxy.pass_through_endpoints.common_utils import encode_bedrock_runtime_modelid_arn
 
 base_llm_http_handler = BaseLLMHTTPHandler()
 from .utils import BasePassthroughUtils
@@ -245,7 +245,7 @@ def llm_passthrough_route(
     
     # need to encode the id of application-inference-profile for bedrock
     if custom_llm_provider == "bedrock" and "application-inference-profile" in endpoint:
-        encoded_url_str = encode_bedrock_runtime_modelid_arn(str(updated_url))
+        encoded_url_str = CommonUtils.encode_bedrock_runtime_modelid_arn(str(updated_url))
         updated_url = httpx.URL(encoded_url_str)
     
     # Add or update query parameters
